# packages/domain-types/openapi.json
{
  "components": {
    "schemas": {
...
      "BuildRecordRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/BuildEntryRead"
            },
            "title": "Entries",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "identifier": {
            "title": "Identifier",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "performed_by": {
            "title": "Performed By",
            "type": "string"
          },
          "performed_on": {
            "format": "date",
            "title": "Performed On",
            "type": "string"
          },
          "procedure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Procedure"
          },
          "prototype": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrototypeSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "string"
          },
          "watch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WatchSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "created_at",
          "created_by",
          "updated_at",
          "updated_by",
          "id",
          "identifier",
          "title",
          "performed_on",
          "performed_by",
          "prototype",
          "watch",
          "procedure",
          "notes",
          "entries"
        ],
        "title": "BuildRecordRead",
        "type": "object"
      },
...
      "BuildRecordUpdate": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "BuildRecordUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PartInstanceRead": {
        "properties": {
          "component": {
            "$ref": "#/components/schemas/ComponentSummary"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "current_prototype": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrototypeSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_watch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WatchSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "heat_treatment_lot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heat Treatment Lot"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "identifier": {
            "title": "Identifier",
            "type": "string"
          },
          "is_placeholder": {
            "title": "Is Placeholder",
            "type": "boolean"
          },
          "lot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot"
          },
          "material_lot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Lot"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "revision": {
            "$ref": "#/components/schemas/RevisionSummary"
          },
          "serial_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number"
          },
          "source": {
            "$ref": "#/components/schemas/PartSource"
          },
          "status": {
            "$ref": "#/components/schemas/PartInstanceStatus"
          },
          "supplier_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Note"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "string"
          }
        },
        "required": [
          "created_at",
          "created_by",
          "updated_at",
          "updated_by",
          "id",
          "identifier",
          "serial_number",
          "status",
          "source",
          "is_placeholder",
          "component",
          "revision",
          "lot",
          "material_lot",
          "heat_treatment_lot",
          "supplier_note",
          "notes",
          "current_prototype",
          "current_watch"
        ],
        "title": "PartInstanceRead",
        "type": "object"
      },
      "PartInstanceStatus": {
        "enum": [
          "AVAILABLE",
          "INSTALLED",
          "REMOVED",
          "SCRAPPED"
        ],
        "title": "PartInstanceStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "\nInternal engineering platform for **NEMETH \u2014 Detroit**: product definition, components\nwith immutable revisions, recursive bills of materials, and (in later slices) prototypes,\nexperiments, measurements and serialized watch genealogy.\n\nErrors are returned as RFC 9457 problem details (`application/problem+json`).\nResources are addressed by UUID or by human identifier (`N1-MVT-002`).\n",
    "title": "NEMETH Engineering Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/builds/{ref}": {
      "get": {
        "operationId": "get_build_api_v1_builds__ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildRecordRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Build",
        "tags": [
          "prototypes"
        ]
      },
      "patch": {
        "operationId": "update_build_api_v1_builds__ref__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildRecordUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildRecordRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Build",
        "tags": [
          "prototypes"
        ]
      }
    },
...
    "/api/v1/components/{ref}/part-instances": {
      "get": {
        "operationId": "component_part_instances_api_v1_components__ref__part_instances_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PartInstanceRead"
                  },
                  "title": "Response Component Part Instances Api V1 Components  Ref  Part Instances Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Component Part Instances",
        "tags": [
          "prototypes"
        ]
      }
    },
...
    "/api/v1/part-instances": {
      "get": {
        "operationId": "list_part_instances_api_v1_part_instances_get",
        "parameters": [
          {
            "description": "Search instance, serial, component",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search instance, serial, component",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PartInstanceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "component_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Component Id"
            }
          },
          {
            "in": "query",
            "name": "prototype_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prototype Id"
            }
          },
          {
            "in": "query",
            "name": "watch_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Watch Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PartInstanceRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Part Instances",
        "tags": [
          "prototypes"
        ]
      },
      "post": {
        "operationId": "create_part_instances_api_v1_part_instances_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartInstanceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PartInstanceRead"
                  },
                  "title": "Response Create Part Instances Api V1 Part Instances Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Part Instances",
        "tags": [
          "prototypes"
        ]
      }
    },
    "/api/v1/part-instances/{ref}": {
      "get": {
        "operationId": "get_part_instance_api_v1_part_instances__ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartInstanceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Part Instance",
        "tags": [
          "prototypes"
        ]
      },
      "patch": {
        "operationId": "update_part_instance_api_v1_part_instances__ref__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartInstanceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartInstanceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Part Instance",
        "tags": [
          "prototypes"
        ]
      }
    },
...
    "/api/v1/prototypes": {
      "get": {
        "operationId": "list_prototypes_api_v1_prototypes_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PrototypeStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PrototypeRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Prototypes",
        "tags": [
          "prototypes"
        ]
      },
      "post": {
        "operationId": "create_prototype_api_v1_prototypes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrototypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototypeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Prototype",
        "tags": [
          "prototypes"
        ]
      }
    },
    "/api/v1/prototypes/{ref}": {
      "get": {
        "operationId": "get_prototype_api_v1_prototypes__ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototypeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Prototype",
        "tags": [
          "prototypes"
        ]
      },
      "patch": {
        "operationId": "update_prototype_api_v1_prototypes__ref__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrototypeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototypeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Prototype",
        "tags": [
          "prototypes"
        ]
      }
    },
    "/api/v1/prototypes/{ref}/builds": {
      "get": {
        "operationId": "list_builds_api_v1_prototypes__ref__builds_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BuildRecordRead"
                  },
                  "title": "Response List Builds Api V1 Prototypes  Ref  Builds Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Builds",
        "tags": [
          "prototypes"
        ]
      },
      "post": {
        "operationId": "create_build_api_v1_prototypes__ref__builds_post",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildRecordRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Build",
        "tags": [
          "prototypes"
        ]
      }
    },
    "/api/v1/prototypes/{ref}/configuration": {
      "get": {
        "operationId": "get_configuration_api_v1_prototypes__ref__configuration_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitConfiguration"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Configuration",
        "tags": [
          "prototypes"
        ]
      }
    },
...
    "/api/v1/revisions/{revision_id}/prototypes": {
      "get": {
        "description": "Prototypes that currently contain a part made to this exact revision.",
        "operationId": "revision_prototypes_api_v1_revisions__revision_id__prototypes_get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrototypeSummary"
                  },
                  "title": "Response Revision Prototypes Api V1 Revisions  Revision Id  Prototypes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revision Prototypes",
        "tags": [
          "prototypes"
        ]
      }
    },
...
  }
}